# api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: MapleOS API
  description: AI Native 多 Agent 协作工作站操作系统 API
  version: 2.1.0
  contact:
    name: MapleOS Team
    url: https://github.com/hongmaple0820/maple-os
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:7788
    description: 本地开发服务器
  - url: https://api.mapleos.dev
    description: 生产服务器
...
paths:
...
  /api/workflows:
    get:
      tags: [workflow]
      summary: 获取工作流列表
      operationId: listWorkflows
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 工作流列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  workflows:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workflow'
...
  /api/agents:
    get:
      tags: [agent]
      summary: 获取Agent列表
      operationId: listAgents
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Agent列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'

    post:
      tags: [agent]
      summary: 注册新Agent
      operationId: registerAgent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  example: my-agent
                description:
                  type: string
                transport_type:
                  type: string
                  enum: [websocket, webhook, mcp, rest, sse]
                  default: websocket
                transport_config:
                  type: object
                capabilities:
                  type: array
                  items:
                    type: string
                max_concurrent_tasks:
                  type: integer
                  default: 3
      responses:
        '200':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          description: 请求参数错误
...
  /api/tasks/stats:
    get:
      tags: [task]
      summary: 获取任务统计
      operationId: getTaskStats
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 任务统计
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStats'
  /api/config:
    get:
      tags: [config]
      summary: 获取系统配置
      operationId: getConfig
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 系统配置
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemConfig'

    put:
      tags: [config]
      summary: 更新系统配置
      operationId: updateConfig
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SystemConfig'
      responses:
        '200':
          description: 更新成功
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Workflow:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        version:
          type: integer
        status:
          type: string
          enum: [draft, active, paused, failed]
        created_at:
          type: integer
          format: int64
        updated_at:
          type: integer
          format: int64
    Agent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [online, busy, offline]
        transport_type:
          type: string
        capabilities:
          type: object
        last_heartbeat:
          type: integer
          format: int64
        max_concurrent_tasks:
          type: integer
...
    TaskStats:
      type: object
      properties:
        total:
          type: integer
        pending:
          type: integer
        running:
          type: integer
        completed:
          type: integer
        failed:
          type: integer
        dead_letter:
          type: integer
    SystemConfig:
      type: object
      properties:
        host:
          type: string
        port:
          type: integer
        database_url:
          type: string
        jwt_secret:
          type: string
        require_auth:
          type: boolean
        admin_username:
          type: string
        admin_password:
          type: string
        usage_limit_usd:
          type: number
        log_level:
          type: string
  # ============================================================
  # v2.1.0 — Execution Fact Chain (Track 1)
  # ============================================================
...